CON-775 Fixed

+++ ws-common/src/main/java/org/intalio/tempo/security/ws/RBACAdminWS.java
@@ -23,6 +23,7 @@ import org.intalio.tempo.security.rbac.RoleNotFoundException;
 import org.intalio.tempo.security.rbac.UserExistsException;
 import org.intalio.tempo.security.rbac.UserNotFoundException;
 import org.intalio.tempo.security.rbac.provider.RBACProvider;
+import org.intalio.tempo.security.simple.SimpleSecurityProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -110,7 +111,17 @@ public class RBACAdminWS extends BaseWS {
                             usersRBACAdmin.setRoleProperties(role, request.getProperties(RBACAdminConstants.DETAILS));
                         }
                     } else if (action.equals(RBACAdminConstants.DELETE_ACTION)) {
-                        usersRBACAdmin.deleteRole(role);
+                        if (!checkRoleAssigned(role, usersRBACProvider)) {
+                            usersRBACAdmin.deleteRole(role);
+                        } else {
+                            RoleExistsException e = new RoleExistsException(
+                                    "Cannot delete role : " + role
+                                            + " is assigned to some user");
+                            LOG.error("Cannot delete role : " + role
+                                    + " is assigned to some user");
+                            throw new Fault(e,
+                                    getRoleExistsExceptionResponse(e));
+                        }
                     } else if (action.equals(RBACAdminConstants.ADD_ACTION)) {
                         RoleExistsException e = new RoleExistsException("Role: " + role + " already exists");
                         LOG.error("Role: " + role + " already exists");
@@ -381,6 +392,22 @@ public class RBACAdminWS extends BaseWS {
         return exists;
     }
 
+    private static boolean checkRoleAssigned(String role,
+            RBACProvider usersRBACProvider) {
+        boolean assigned = true;
+        try {
+            if (_securityProvider instanceof SimpleSecurityProvider) {
+                String[] roles = usersRBACProvider.getQuery().assignedUsers(
+                        role);
+                if (roles == null || roles.length == 0) {
+                    assigned = false;
+                }
+            }
+        } catch (Exception e) {
+            assigned = false;
+        }
+        return assigned;
+    }
     private static OMElement elementProperty(String name, String Value) {
         OMElement prop = element(RBACAdminConstants.PROPERTY);
         prop.addChild(elementText(RBACAdminConstants.NAME, name));
